+++ lib/devise/orm/mongoid.rb
@@ -1,3 +1,5 @@
-require 'orm_adapter/adapters/mongoid'
+ActiveSupport.on_load(:mongoid) do
+  require 'orm_adapter/adapters/mongoid'
 
-Mongoid::Document::ClassMethods.send :include, Devise::Models
+  Mongoid::Document::ClassMethods.send :include, Devise::Models
+end